Add Photos directory, work on Hero section
Started working on the Hero section of the webpage, and
added a photos directory for all the photos that will be
eventually added to the page. Currently encountering spacing issues,
and will have to do some research on how to get the CSS to play nicely
with all the various elements on the page.