Add the remaining page information. Adjust CSS
The rest of the information that was going onto the page was added,
as well as the information that I wanted to change was modified to
say what I wanted it to say. At this point, the only remaining step in
the project is to adjust the CSS, formatting, and see if I can clean
up the code a little.